requirements.txt
yfinance
pandas
schedule
pyTelegramBotAPI
python-dotenv
